Compute the contents of MANIFEST.MF
kotlinx_serialization_core is supposed to have Implementation-Version and
Require-Kotlin-Version fields in its MANIFEST.MF, which the compiler
plugin will use to verify that a valid version of the runtime is present.
Read the kotlinx_serialization_core version out of METADATA and the
required kotlin version out of core/build_gradle.kts to generate
the manifest.

Also add kotlinx_serialization_core_tests as a test that the compiler
plugin works.

Test: kotlinx_serialization_core_tests
Bug: 363909427
Change-Id: I77147a64584a0cc3a8abd03897628d23693bda2e